Removed buildAnnotatedString method.
