+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/detail/RestaurantDetail.kt
@@ -1,6 +1,5 @@
 package com.jordangellatly.hermosahappyhour.ui.detail
 
-import android.os.CountDownTimer
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
@@ -9,8 +8,6 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.*
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -18,13 +15,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.jordangellatly.hermosahappyhour.R
-import com.jordangellatly.hermosahappyhour.model.Event
 import com.jordangellatly.hermosahappyhour.model.EventType
 import com.jordangellatly.hermosahappyhour.model.Restaurant
 import com.jordangellatly.hermosahappyhour.model.RestaurantRepo
@@ -71,7 +64,8 @@ fun RestaurantDetail(
                 )
             }
             restaurant.eventsByDate.getValue(formattedDateTimestamp)[EventType.HappyHour]?.let { event ->
-                val weeklyHours = getWeeklyEventScheduleFromRestaurant(restaurant, EventType.HappyHour)
+                val weeklyHours =
+                    getWeeklyEventScheduleFromRestaurant(restaurant, EventType.HappyHour)
                 HappyHour(
                     weeklyHours = weeklyHours,
                     eventStart = event.startTimestamp,
@@ -89,7 +83,8 @@ fun RestaurantDetail(
                 )
             }
             restaurant.eventsByDate.getValue(formattedDateTimestamp)[EventType.Special]?.let { event ->
-                val weeklyHours = getWeeklyEventScheduleFromRestaurant(restaurant, EventType.Special)
+                val weeklyHours =
+                    getWeeklyEventScheduleFromRestaurant(restaurant, EventType.Special)
                 SpecialEvent(
                     weeklyHours = weeklyHours,
                     eventStart = event.startTimestamp,
@@ -101,83 +96,6 @@ fun RestaurantDetail(
         }
     }
 }
-
-@Composable
-fun buildAnnotatedTimerString(event: Event) =
-    buildAnnotatedString {
-        var millisInFuture = 0L
-        var timeIndicatorColor = HermosaHappyHourTheme.colors.textSecondary
-        val currentTime = Calendar.getInstance()
-        val timestampFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm", Locale.getDefault())
-        val startDate = timestampFormat.parse(event.startTimestamp)
-        val startTime = Calendar.getInstance().apply {
-            if (startDate != null) {
-                time = startDate
-            }
-        }
-        val endDate = timestampFormat.parse(event.endTimestamp)
-        val endTime = Calendar.getInstance().apply {
-            if (endDate != null) {
-                time = endDate
-            }
-        }
-        val stringStart = formatTimestamp(event.startTimestamp, "ha")
-        val stringEnd = formatTimestamp(event.endTimestamp, "ha")
-        when {
-            currentTime < startTime -> {
-                timeIndicatorColor = Color.Green
-                withStyle(style = SpanStyle(timeIndicatorColor)) {
-                    append("Starts")
-                }
-                append(" at $stringStart")
-                millisInFuture = startTime.timeInMillis - currentTime.timeInMillis
-            }
-            currentTime > startTime && currentTime < endTime -> {
-                timeIndicatorColor = HermosaHappyHourTheme.colors.orange
-                withStyle(style = SpanStyle(timeIndicatorColor)) {
-                    append("Ends")
-                }
-                append(" at $stringEnd")
-                millisInFuture = endTime.timeInMillis - currentTime.timeInMillis
-            }
-            currentTime > endTime -> {
-                timeIndicatorColor = Color.Red
-                withStyle(style = SpanStyle(timeIndicatorColor)) {
-                    append("Ended")
-                }
-                append(" at $stringEnd")
-                millisInFuture = 0L
-            }
-            else -> {
-                append("")
-            }
-        }
-        append(" \u2022 ")
-
-        val timeData = remember { mutableStateOf(millisInFuture) }
-        val countDownTimer =
-            object : CountDownTimer(millisInFuture, 1000) {
-                override fun onTick(millisUntilFinished: Long) {
-                    timeData.value = millisUntilFinished
-                }
-
-                override fun onFinish() {}
-            }
-
-        DisposableEffect(key1 = "key") {
-            countDownTimer.start()
-            onDispose {
-                countDownTimer.cancel()
-            }
-        }
-
-        val offset = SimpleDateFormat("HH:mm:ss", Locale.US)
-        offset.timeZone = TimeZone.getTimeZone("GMT")
-        val timeText = offset.format(timeData.value)
-        withStyle(style = SpanStyle(timeIndicatorColor)) {
-            append(timeText)
-        }
-    }
 
 fun getWeeklyEventScheduleFromRestaurant(restaurant: Restaurant, eventType: EventType) =
     restaurant.eventsByDate
